mfd: intel-m10-bmc: create max10 bcm security engine
Create a platform driver that can be invoked as a sub
driver for the Intel MAX10 BMC in order to support
secure updates. This sub-driver will invoke an
instance of the FPGA security manager in order
to expose sysfs interfaces for managing and monitoring
secure updates to FPGA and BMC images and firmware.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>