fpga: dfl split input check in dfl_fpga_set_irq_triggers
The overflow check in the compound check is unlikely.

So split the check and wrap the overflow check with the unlikely macro.

The second check to validate the index is not larger the the number irqs
is repeated in the do_set_irq_trigger.

Because do_set_irq_trigger is called in a loop, remove the redunant check
in do_set_ieq_trigger.

Signed-off-by: Tom Rix <trix@redhat.com>