spi: altera: support indirect access to the registers
This patch adds support for indirect access to the registers via parent
regmap.

The use case is, the spi master is a sub device of an Multifunction
device, which is connected to host by some indirect bus. To support this
device type, a new platform_device_id is introduced, and the driver tries
to get parent regmap for it like many MFD sub device drivers do.

The new added platform_device_id "dfl_spi_altera" is for the spi master
implemented under FPGA Device Feature List (DFL) framework. The DFL
based FPGA card uses a dedicated indirect bus to access this spi master.

Signed-off-by: Xu Yilun <yilun.xu@intel.com>