media: dvb-frontends/cxd2099: report errors
Clang static analysis reports this error

drivers/media/dvb-frontends/cxd2099.c:420:2: warning: Undefined
  or garbage value returned to caller
        return val;
        ^~~~~~~~~~

In read_cam_control, the call to read_io can fail.
When it fails val is not set.

The failure status should be returned to the caller,
not the unset val.

Similar problem with read_attribute_mem

Fixes: 0f0b270f905b ("[media] ngene: CXD2099AR Common Interface driver")

Signed-off-by: Tom Rix <trix@redhat.com>