misc: rtsx: improve status check
clang static analysis flags this error

rtsx_usb.c:505:10: warning: The left operand of '&'
  is a garbage value
        if (val & cd_mask[card])
            ~~~ ^

val is set when rtsx_usb_get_card_status() is successful.  The
problem is how it checks its callers returns.

	/* usb_control_msg may return positive when success */
	if (ret < 0)
		return ret;

This is correct for the usb_control_msg() the call. However,
the call to rtsx_usb_get_status_with_bulk is only successful
when 0 is returned.

So make status checking block specific.

Fixes: 730876be2566 ("mfd: Add realtek USB card reader driver")
Signed-off-by: Tom Rix <trix@redhat.com>