bus: fsl-mc: fix improper free of mc_dev
Clang static analysis reports this error

fsl-mc-bus.c:891:2: warning: Attempt to free released memory
        kfree(mc_dev);
        ^~~~~~~~~~~~~

In this block of code

if (strcmp(obj_desc->type, "dprc") == 0) {
..
  mc_bus = kzalloc(..)
  mc_dev = &mc_bus->mc_dev;

mc_dev is not alloc-ed, so it should not be freed.
Old handler triggers a false positive from checkpatch, so add a
comment and change logic a bit.

Fixes: a042fbed0290 ("staging: fsl-mc: simplify couple of deallocations")
Signed-off-by: Tom Rix <trix@redhat.com>