media: imx: imx7_mipi_csis: convert some switch cases to the default
Static analysis reports this false positive
imx7-mipi-csis.c:1027:2: warning: 4th function call argument is
  an uninitialized value

The variable 'align' is falsly reported as uninitialized.
Even though all the cases are covered in the
	switch (csis_fmt->width % 8) {

Because there is no default case, it is reported as uninialized.

In general all switches should have a default and this case can
be improved be converting the most numerous set of cases to the
default.

Signed-off-by: Tom Rix <trix@redhat.com>