treewide: Add missing semicolons to __assign_str uses
The __assign_str macro has an unusual ending semicolon but the vast
majority of uses of the macro already have semicolon termination.

$ git grep -P '\b__assign_str\b' | wc -l
551
$ git grep -P '\b__assign_str\b.*;' | wc -l
480

Add semicolons to the __assign_str() uses without semicolon termination
and all the other uses without semicolon termination via additional defines
that are equivalent to __assign_str() with the eventual goal of removing
the semicolon from the __assign_str() macro definition.

Link: https://lore.kernel.org/lkml/1e068d21106bb6db05b735b4916bb420e6c9842a.camel@perches.com/
Link: https://lkml.kernel.org/r/144460ce4f34a51dabb76e422a718573db77cdc8.camel@perches.com

Signed-off-by: Joe Perches <joe@perches.com>
Acked-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>