of: remove duplicate declaration of of_iomap()
A ranconfig produces this linker error
irq-al-fic.c:252: undefined reference to `of_iomap'

The declaration of of_iomap() is dependent on OF
The definition of of_iomap() is dependent on OF_ADDRESS
These should match.  There are duplicate declarations
of of_iomap(), remove of_iomap() and the
of_address_to_resource() duplicate.

Signed-off-by: Tom Rix <trix@redhat.com>